+++ src/main/java/de/fraunhofer/iosb/ilt/frostclient/models/swecommon/simple/CountRange.java
@@ -104,7 +104,7 @@ public class CountRange extends AbstractRange<CountRange, Long> {
     @Override
     public boolean validate(List<Long> input) {
         if (input == null) {
-            return isOptional();
+            return isOptional() || isSecret();
         }
         final int size = input.size();
         if (size != 2) {