+++ src/main/java/de/fraunhofer/iosb/ilt/frostclient/models/swecommon/simple/Text.java
@@ -107,7 +107,7 @@ public class Text extends AbstractSimpleComponent<Text, String> {
     @Override
     public boolean validate(Object input) {
         if (input == null) {
-            return isOptional();
+            return isOptional() || isSecret();
         }
         if (input instanceof JsonNode j) {
             return validate(j);
@@ -122,7 +122,7 @@ public class Text extends AbstractSimpleComponent<Text, String> {
     @Override
     public boolean validate(JsonNode input) {
         if (input == null) {
-            return isOptional();
+            return isOptional() || isSecret();
         }
         if (!input.isTextual()) {
             LOGGER.debug("Non-Text value {} for Text.", input);
@@ -133,7 +133,7 @@ public class Text extends AbstractSimpleComponent<Text, String> {
 
     public boolean validate(String input) {
         if (input == null) {
-            return isOptional();
+            return isOptional() || isSecret();
         }
         if (constraint == null) {
             return true;