+++ src/main/java/de/fraunhofer/iosb/ilt/frostclient/models/swecommon/simple/TimeRange.java
@@ -161,7 +161,7 @@ public class TimeRange extends AbstractRange<TimeRange, String> {
     @Override
     public boolean validate(List<String> input) {
         if (input == null) {
-            return isOptional();
+            return isOptional() || isSecret();
         }
         final int size = input.size();
         if (size != 2) {